Added gem 'hydra-head' to Gemfile
I found that gem 'hydra-head' had to be in the Gemfile in order for hydra-jetty to start successfully. Without hydra-head in the Gemfile, I would get "Unable to move tmp/jetty_generator/hydra-jetty-master into jetty/ Operation not permitted - (tmp/jetty_generator/hydra-jetty-master, jetty)" when I tried to run the hydra-jetty generator. With hydra-head in the Gemfile, this problem was solved. I tried this after seeing the instructions for hydra-head installation at https://github.com/projecthydra/hydra-head.